Testing and fixed some issues
